Move build_pyrenew_model to pyrenew_hew
