+++ pipelines/tests/test_build_pyrenew_model.py
@@ -4,8 +4,8 @@ import jax.numpy as jnp
 import polars as pl
 import pytest
 
-from pipelines.build_pyrenew_model import build_pyrenew_model
 from pipelines.utils import get_model_data_and_priors_from_dir
+from pyrenew_hew.util import build_pyrenew_model
 
 
 @pytest.fixture
